AdvancedAgents/McK-Consutlancy/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
anthropic>=0.40.0
sqlalchemy==2.0.36
python-docx==1.1.2
markdown==3.7
python-dotenv==1.0.1
aiofiles==24.1.0
python-multipart==0.0.12
pydantic>=2.9.0
httpx>=0.27.0
xhtml2pdf>=0.2.11
htmldocx>=0.0.6
Pillow>=10.0.0
beautifulsoup4>=4.12.0
openpyxl>=3.1.0
